vk: extract more stuff from framectl
- extract render pass into vk_render_pass.{c,h}
- move rendering/render pass setup into vk_render.c
- move swapchain init into vk_core.c
